% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_consistency}
\alias{check_consistency}
\title{Check the consistence of the nodes, flows and palette data.frames}
\usage{
check_consistency(nodes, flows, palette = NULL)
}
\arguments{
\item{nodes}{data.frame containing the nodes definition}

\item{flows}{data.frame containing the flows definition}

\item{palette}{data.frame containing the palette definition}
}
\value{
TRUE if all checks are passed; FALSE otherwise.
}
\description{
Check the consistence of the nodes, flows and palette data.frames
}
\examples{
nodes <- data.frame(ID=c("A","B"), x=1:2, y=0)
flows <- data.frame(from="A", to="B", quantity=10)
check_consistency(nodes, flows)
}
