% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFA.R
\docType{data}
\name{MFA}
\alias{MFA}
\title{Material Flow Account sample data}
\format{
A list with three data frames:
\itemize{
\item nodes
\item flows
\item colors
}
}
\usage{
MFA
}
\description{
Dataset containing sample material flow account data,
formatted for use within 'PantaRhei'
}
\keyword{datasets}
