% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{parse_nodes}
\alias{parse_nodes}
\title{Parse the information from a 'nodes' definition table.}
\usage{
parse_nodes(nodes, verbose = FALSE)
}
\arguments{
\item{nodes}{data.frame containing the nodes definition}

\item{verbose}{logical: print some information?}
}
\value{
modified nodes data.frame
}
\description{
This function checks the content of a nodes definition,
and appends some missing columns.
It is mainly used internally, but can be invoked by the uses to see what it does.
}
\examples{
n0 <- data.frame(ID=c("A","B"), x=1:2, y=0)
str(n0)
n1 <- parse_nodes(n0)
str(n1)
}
