% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{parse_palette}
\alias{parse_palette}
\title{Parse the information from a 'palette' definition table.}
\usage{
parse_palette(palette, verbose = TRUE)
}
\arguments{
\item{palette}{data.frame containing the palette definition}

\item{verbose}{logical: print some information?}
}
\value{
modified palette data.frame
}
\description{
This function checks the content of a palette definition,
and appends some missing columns.
It is mainly used internally, but can be invoked by the uses to see what it does.
}
\examples{
p0 <- data.frame(substance="any", color="red")
str(p0)
p1 <- parse_palette(p0)
str(p1) # Should be the same!
}
