% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{strformat}
\alias{strformat}
\title{Format a string}
\usage{
strformat(s, ...)
}
\arguments{
\item{s}{character string to be formatted}

\item{...}{formatting specifiers to be forwarded to gpar()}
}
\value{
formatted string
}
\description{
This function adds formatting information to a character string by storing
this information as the character string's attributes.
Run the example to see how it works.
}
\details{
All formatting specifiers work as if \code{gpar()} would be called.
(It is, behind the screen.)
}
\examples{
s <- strformat("Hello, World", fontsize=18, col="red")
str(s)  # show object structure
}
